refactor: 1. replace `isEmailValid` function 2. Make changes in `ImpersonateUserForm`
